Added a bit of threading
Have much more work to do but did a bit of performance work to make the thing run a bit quicker when being run for the first time. Python defaults to one thread and a lot of the work that is done in this could easily be done in parallel so I've started to add that to the various programs.